+++ revisit-alerts-lint/src/main/java/com/lachlanmckee/revisit/date/Month.kt
@@ -1,0 +1,5 @@
+package com.lachlanmckee.revisit.date
+
+enum class Month {
+  JANUARY, FEBRUARY, MARCH, APRIL, MAY, JUNE, JULY, AUGUST, SEPTEMBER, OCTOBER, NOVEMBER, DECEMBER
+}