+++ revisit-alerts-sample/src/main/java/com/lachlanmckee/revisit/android/RevisitByDateExample.kt
@@ -1,9 +1,10 @@
 package com.lachlanmckee.revisit.android
 
 import android.annotation.SuppressLint
+import com.lachlanmckee.revisit.Month
 import com.lachlanmckee.revisit.RevisitByDate
 
 // Remove suppress to see demonstrate lint rule.
 @SuppressLint("RevisitDate")
-@RevisitByDate(year = 1999, month = 1, day = 1, reason = "For testing purposes")
+@RevisitByDate(day = 1, month = Month.JANUARY, year = 1999, reason = "For testing purposes")
 class RevisitByDateExample