+++ CHANGELOG.md
@@ -4,4 +4,4 @@ Change Log
 Version 1.0.0 *(2021-08-29)*
 ----------------------------
 
-* Initial release with support for `RevisitDate`
+* Initial release with support for `RevisitByDate` and `RevisitFromDate`